v1.7: liveSRRF and v1.0 ParamSweep. Lots of tidying up, drop down menu for liveSRRF temporal analysis, pixel width and units for Sweep maps.

+++ src/nanoj/liveSRRF/ErrorMapLiveSRRF.java
@@ -319,7 +319,7 @@ public class ErrorMapLiveSRRF {
     private FloatProcessor getRSF(float sigma) {
 
         // calculate the final RSF
-        float sigma2 = (float) (ROOT2 * abs(sigma));
+        float sigma2 = (ROOT2 * abs(sigma));
         int radius = max(((int) sigma) * 3, 1);
         int size = radius * 2 + 1;
         float vKernelSum = 0;
